pcstac/requirements-server.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --extra=server --output-file=pcstac/requirements-server.txt ./pcstac/setup.py
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
    # via stac-fastapi-pgstac
attrs==23.2.0
    # via
    #   stac-fastapi-pgstac
    #   stac-fastapi-types
brotli==1.1.0
    # via brotli-asgi
brotli-asgi==1.4.0
    # via
    #   stac-fastapi-api
    #   stac-fastapi-pgstac
buildpg==0.4
    # via stac-fastapi-pgstac
cachetools==5.3.3
    # via pypgstac
click==8.1.7
    # via
    #   stac-pydantic
    #   uvicorn
dateparser==1.2.0
    # via pygeofilter
exceptiongroup==1.2.1
    # via anyio
fastapi-slim==0.111.0
    # via stac-fastapi-types
fire==0.4.0
    # via pypgstac
geojson-pydantic==1.1.0
    # via stac-pydantic
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.7
    # via
    #   anyio
    #   pcstac (pcstac/setup.py)
iso8601==2.1.0
    # via stac-fastapi-types
lark==0.12.0
    # via pygeofilter
orjson==3.10.4
    # via
    #   pcstac (pcstac/setup.py)
    #   pypgstac
    #   stac-fastapi-pgstac
plpygis==0.2.2
    # via pypgstac
psycopg[binary]==3.1.18
    # via pypgstac
psycopg-binary==3.1.18
    # via psycopg
psycopg-pool==3.1.9
    # via pypgstac
pydantic==2.7.4
    # via
    #   fastapi-slim
    #   geojson-pydantic
    #   pydantic-settings
    #   pypgstac
    #   stac-fastapi-pgstac
    #   stac-pydantic
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.3.3
    # via stac-fastapi-types
pygeofilter==0.2.1
    # via stac-fastapi-pgstac
pygeoif==1.5.0
    # via pygeofilter
pypgstac[psycopg]==0.8.6
    # via
    #   pcstac (pcstac/setup.py)
    #   stac-fastapi-pgstac
pystac==1.10.1
    # via pcstac (pcstac/setup.py)
python-dateutil==2.8.2
    # via
    #   dateparser
    #   pypgstac
    #   pystac
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
pytz==2024.1
    # via dateparser
pyyaml==6.0.1
    # via uvicorn
regex==2024.5.15
    # via dateparser
six==1.16.0
    # via
    #   fire
    #   python-dateutil
smart-open==7.0.4
    # via pypgstac
sniffio==1.3.1
    # via anyio
stac-fastapi-api==3.0.0a3
    # via
    #   pcstac (pcstac/setup.py)
    #   stac-fastapi-extensions
    #   stac-fastapi-pgstac
stac-fastapi-extensions==3.0.0a3
    # via
    #   pcstac (pcstac/setup.py)
    #   stac-fastapi-pgstac
stac-fastapi-pgstac==3.0.0a1
    # via pcstac (pcstac/setup.py)
stac-fastapi-types==3.0.0a3
    # via
    #   pcstac (pcstac/setup.py)
    #   stac-fastapi-api
    #   stac-fastapi-extensions
    #   stac-fastapi-pgstac
stac-pydantic==3.1.0
    # via
    #   stac-fastapi-pgstac
    #   stac-fastapi-types
starlette==0.37.2
    # via
    #   brotli-asgi
    #   fastapi-slim
tenacity==8.1.0
    # via pypgstac
termcolor==2.4.0
    # via fire
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi-slim
    #   pcstac (pcstac/setup.py)
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   pygeoif
    #   starlette
    #   uvicorn
tzlocal==5.2
    # via dateparser
uvicorn[standard]==0.30.1
    # via pcstac (pcstac/setup.py)
uvloop==0.19.0
    # via uvicorn
version-parser==1.0.1
    # via pypgstac
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
wrapt==1.16.0
    # via smart-open
